Fix tmux helper script to work in zsh
This script was broken because `zsh` handles variable interpolation in
scripts differently, treating the whole string as a single argument
rather than splitting it up by spaces like `bash` does.

This resulted in us having to "un-DRY" the code a bit in order to get
it cross-compatible with both shells.