Slugify session names to not have periods
As of tmux 1.9a, periods are not allowed in session names. Thus we must
replace them with an allowed character (hyphen, in this case) in order
to continue allowing this script to work with certain directories.

While this does allowed for two differently-named directories to have
the same session name (e.g. `.dotfiles` and `-dotfiles`), this is pretty
unlikely to be an issue.