Downgrade tmux config to be 1.8 specific
This reverts commit 787b4061737725deeec4d92b401b413750e1b9e6.

Since CentOS 7 ships with tmux 1.8, and I don't want to have to always
install tmux 1.9a every time I want my configuration to work, I'm
intentionally switching the configuration to be 1.8 specific.

Sadly, tmux does not provide a way to allow you to open panes in the
same directly in a way that is compatible with both 1.8 and 1.9, so I'm
going to stick with 1.8 so I get this for free on every CentOS 7 system
I use.