Add support for shell-specific plugin install scripts
While we have support for setting up plugins based on the shell, we
didn't have support for custom installation based on the shell (i.e. we
would always run `setup.sh`, rather than looking for a `setup.zsh` or
similar to see if any custom installation was required.

This commit adds support for these specific installation scripts.

As usual, we will always run `.sh` setup scripts before the specific
script.

This required some additional changes to the `install` script, as we
that previously always invoked the shell within `bash`. Now it instead
calls the new `bootstrap/install.sh` script with the user's current
`$SHELL`, ensuring we always run the code within the intended shell,
allowing the shell detection to work.