Add support for zsh syntax highlighting/completion
This adds three projects which will make using zsh incredibly
convenient.

The syntax highlighting component allows zsh to highlight your commands
as you type, similar to the `fish` shell.

The completions allow you to easily hit tab to see a list of possible
completions to a command. zsh comes with a bunch by default, but the
separate `zsh-completions` repo gives us a bunch extra that haven't been
merged upstream yet. In the process of adding this, we removed some
completion files as they are replaced by this repo.

Lastly, the `zsh-history-substring-search` project provides a convenient
shortcut for cycling through old commands that match the prefix of the
command you've already typed.

Lastly, we customize zsh's command line to act like the bash shell we're
used to by setting it to emacs mode.