Symlink .inputrc instead of setting INPUTRC
Previously, we set the configuration file for `readline` by setting the
`INPUTRC` variable because we thought symlinks were messy (in that they
added a bunch of files to your home directly.

It's now apparent that it's kind of nice to know what exactly Dot is
installing, so symlinks are preferable to environment variables wherever
possible as they are easier to discover.