Allow plugins to define code to run after all plugins loaded
Sometimes it's useful for a plugin to execute code after all other
plugins have loaded. In zsh for example, we want to be able to use the
same completions for shortcut functions like `t` (for `tmux`), but in
order to do this we need to run `compdef t=tmux` after the completions
for tmux have already been loaded (e.g. `compinit` has been called).

Add support for executing a `finish.*sh` file in each plugin. This will
execute after `bootstrap/finish.*sh`, achieving the behavior we want.