Show intraline differences using diff-highlight
Git ships with a Perl script called `diff-highlight` which displays the
intraline differences in a diff, making it more readable.

Unfortunately, the script is not installed in a directory in your PATH
by default, so you need to add it manually. Furthermore, the script that
ships with the git RPM for CentOS doesn't include executable permissions
by default.

To get around this and make this change more portable, include a copy of
the script in my dotfiles so I don't need to manually set this up each
time depending on the OS I'm running.

Thanks to @lencioni for making the existence of `diff-highlight` known
to me. https://github.com/lencioni/dotfiles/commit/69b47285