Don't confirm Homebrew installation
When you run the Homebrew setup script it asks you to confirm before
continuing:

    Press RETURN to continue or any other key to abort

Since we want to have this process fully automated, pass empty lines to
STDIN so that the confirmation is automatically skipped.

This will allow us to test the install in future as part of a CI run or
similar.