Remove accessibility helpers
As of macOS Sierra the SQLite file is protected by SIP (System Integrity
Protection) which can't be disabled without a system restart. This makes
it impractical to programatically enable/disable accessibility support
for an application.

For now, we'll have to do this manually. Remove the helpers to remove
any confusion.