Switch vim to load .vim/vimrc
As of Vim 7.3.1178 Vim now loads ~/.vim/vimrc if .vimrc is not present,
see: https://github.com/vim/vim/commit/6a459902592e2a4ba68

We can reduce the number of symlinks we have.

While here I also renamed the repo to `vim` so it's no longer a hidden
directory in the plugin.