Fix `rbenv shell` command
Since we were only running the `eval $(rbenv init -)` initializer in a
login shell context, this command would not work inside shells opened by
tmux panes, since we've explicitly configured those to _not_ start with
a login context, but a regular interactive context.

We didn't load them in an interactive context because it was pretty
heavyweight--it would result in us adding shims to the PATH twice (once
in the login shell, and then again in interactive shells in tmux). It
also slowed startup time due to loading completions and running a
rehash.

The workaround is to extract the subset of commands from `rbenv init -`
that we really want to run (namely the rehash/shell commands) and run
nothing else. This accomplishes that.

A final note: I chose to have this run only on interactive shells
rather than all shells because I currently can't think of any use cases
where I would want the former, and even if we did we would need to
figure out how to load the login context commands in this context as
well. Apply YAGNI until proven otherwise for now.